+++ src/engine.ts
@@ -1,5 +1,5 @@
 import { readFile } from "node:fs/promises"
-import { processCVE } from "./cve/processor"
+import { processCVE } from "./engine/cve/processor"
 
 interface CveBoxConfig {
   cves: string[]
