+++ src/engine/ab/archive-box.ts
@@ -1,4 +1,4 @@
-import { exec } from "../utils/exec";
+import { exec } from "../../utils/exec";
 
 export async function archiveURLs(dir: string, urls: string[]) {
   try {
