+++ src/engine/cve/cve.types.ts
@@ -1,4 +1,4 @@
-import { ArchiveBoxIndex } from "../types/ab"
+import { ArchiveBoxIndex } from "../../types/ab"
 
 export interface ArchiveModel {
   archive: ArchiveInstance
