+++ src/engine/cve/processor.ts
@@ -2,13 +2,13 @@ import os from 'node:os'
 import fs from 'node:fs/promises';
 import path from 'node:path';
 import crypto from 'node:crypto'
-import { ArchiveBoxIndex } from '../types/ab';
+import { ArchiveBoxIndex } from '../../types/ab';
 import { encodeRequestToS3Key, upload } from '../aws/s3';
-import { CveBoxDB } from '../data/db';
+import { CveBoxDB } from '../../data/db';
 import { UPSERT_ASSET, UPSERT_REQUEST } from './sql';
 import { archiveURLs } from '../ab/archive-box';
 import { fetchReferences } from './api';
-import { exec } from '../utils/exec';
+import { exec } from '../../utils/exec';
 import { ArchiveModel } from './cve.types';
 
 const DEST_PATH = process.env.COPY_DEST_PATH || "/host_os/mount/archive/"
