+++ src/main.ts
@@ -1,15 +1,16 @@
 import { runEngine } from "./engine";
-import { init } from "./scheduling/scheduler";
+import { runScheduler } from "./scheduler";
+
 
 
 async function main() {
   const [, , mode] = process.argv
   switch (mode) {
     case 'engine':
-      return runEngine()
+      return await runEngine()
 
     case 'scheduler':
-      return await init()
+      return await runScheduler()
 
     default:
       throw new Error(`Invalid mode (${mode || '<undefined>'}) – invoke via "node main.js <mode>`)
