+++ src/scheduler.ts
@@ -1,0 +1,5 @@
+import { init } from "./scheduling/scheduler";
+
+export async function runScheduler() {
+  await init()
+}
